+++ tests/models/test_run_tournaments.py
@@ -8,205 +8,184 @@ from chess.models.rounds import Round
 from chess.models.tournaments import Tournament
 
 
-@pytest.fixture
-def four_players():
-    """load 4 players"""
-
-    p_list = Player.read_all()
-    p_list = [i for i in p_list if i.firstname.startswith("Test")]
-    p_list = p_list[:4]
-
-    assert len(p_list) >= 4
-    assert isinstance(p_list[0], Player)
-    return p_list
-
-
-@pytest.fixture
-def created_default_tournament():
-    """create a tournament"""
-
-    name = "TestTournament" + secrets.token_hex(4)
-    start_date = "2023-01-01"
-    end_date = "2023-12-31"
-
-    t = Tournament(name, start_date, end_date)
-    t.create()
-
-    return t
-
-
-@pytest.fixture
-def loaded_default_tournament():
-    """load a tournament"""
-
-    tournament_list = Tournament.read_all()
-
-    tournament_list = [
-        i for i in tournament_list if i.name.startswith("TestTournament")
-    ]
-    tournament = tournament_list[-1]
-
-    return tournament
-
-
 class TestTournamentRun:
     """Test Tournament model"""
 
-    def test_add_players(self, four_players, created_default_tournament):
+    def test_create_player(self, new_four_players):
+
+        for p in new_four_players:
+            assert isinstance(p, Player)
+
+    def test_add_players(self, four_players, default_tournament):
         """test add 4 players"""
 
         # create 4 players
         logging.info(four_players)
 
         # create tournament
-        logging.info(created_default_tournament)
+        logging.info(default_tournament)
 
-        created_default_tournament.player_id_list = []
+        default_tournament.player_id_list = []
 
         # add players to tournament
         for player in four_players:
-            created_default_tournament.add_player(player.player_id)
+            default_tournament.add_player(player.player_id)
 
-        assert created_default_tournament.n_players == 4
+        assert default_tournament.n_players == 4
 
-    def test_update_status(self, loaded_default_tournament):
+        # just reload the tournament from db to check update OK in db
+        id_tournament = default_tournament.tournament_id
+        tournament = Tournament.search(id_tournament)
+        assert tournament.n_players == 4
+
+    def test_update_status(self, last_tournament):
         """update tournament status to "In Progress"""
 
-        assert loaded_default_tournament.status == "Created"
-        assert loaded_default_tournament.n_players == 4
+        assert last_tournament.status == "Created"
+        assert last_tournament.n_players == 4
 
-        loaded_default_tournament.update_status("In Progress")
+        last_tournament.update_status("In Progress")
 
-    def test_round_computed(self, loaded_default_tournament):
+    def test_round_computed(self, last_tournament):
         """test if rounds had been computed"""
 
-        logging.warning(f"id tournament: {loaded_default_tournament.tournament_id}")
+        logging.warning(f"id tournament: {last_tournament.tournament_id}")
 
-        assert loaded_default_tournament.status == "In Progress"
+        assert last_tournament.status == "In Progress"
 
-        assert len(loaded_default_tournament.round_id_list) == 3
+        assert len(last_tournament.round_id_list) == 3
 
-        # assert loaded_default_tournament.round == 0
+        assert last_tournament.current_round_number == 0
 
-    def test_get_results(self, four_players, loaded_default_tournament):
-        """Test access to results"""
+        logging.critical(
+            f"last_tournament.round_id_list is {last_tournament.round_id_list}"
+        )
+        round_objects = [Round.search(i) for i in last_tournament.round_id_list]
 
-        # Get current round
-        current_round = loaded_default_tournament.get_current_round()
+        id_tournament = last_tournament.tournament_id
 
-        # Check if current_round is not None
-        assert current_round is not None
+        for round_ in round_objects:
+            assert id_tournament in round_.round_id
 
-        # Access match_list directly on the current_round instance
-        assert isinstance(current_round.matches, list)
+    # def test_get_results(self, four_players, last_tournament):
+    #     """Test access to results"""
 
-        # Get 1st match of the round
-        first_match_of_the_list = current_round.matches[0]
-        assert isinstance(first_match_of_the_list, list)
+    #     # Get current round
+    #     current_round = last_tournament.get_current_round()
 
-        # Get 1st player of the 1st match
-        first_player_of_first_match = first_match_of_the_list[0]
-        assert isinstance(first_player_of_first_match, list)
+    #     # Check if current_round is not None
+    #     assert current_round is not None
 
-        # Extract id and score for this player
-        first_player_of_first_match_id = first_player_of_first_match[0]
-        first_player_of_first_match_score = first_player_of_first_match[1]
+    #     # Access match_list directly on the current_round instance
+    #     assert isinstance(current_round.matches, list)
 
-        # Check values
-        assert first_player_of_first_match_id == four_players[0].player_id
-        assert first_player_of_first_match_score in [0, 1]
+    #     # Get 1st match of the round
+    #     first_match_of_the_list = current_round.matches[0]
+    #     assert isinstance(first_match_of_the_list, list)
 
-    def test_add_1st_results(self, four_players, loaded_default_tournament):
-        """add results for round 1"""
+    #     # Get 1st player of the 1st match
+    #     first_player_of_first_match = first_match_of_the_list[0]
+    #     assert isinstance(first_player_of_first_match, list)
 
-        # store current_round_number
-        current_round_number = loaded_default_tournament.current_round_number
+    #     # Extract id and score for this player
+    #     first_player_of_first_match_id = first_player_of_first_match[0]
+    #     first_player_of_first_match_score = first_player_of_first_match[1]
 
-        # use random results for the first round
-        new_res = [
-            # 1st match
-            (
-                # player 1 of 1st match
-                (four_players[0].player_id, 1),  # id, score
-                # player 2 of 1st match
-                (four_players[1].player_id, 0),  # id, score
-            ),
-            # 2nd match
-            (
-                # 1st player of 2nd match
-                (four_players[2].player_id, 1),  # id, score
-                # 2nd player of the 2nd match
-                (four_players[3].player_id, 0),  # id, score
-            ),
-        ]
+    #     # Check values
+    #     assert first_player_of_first_match_id == four_players[0].player_id
+    #     assert first_player_of_first_match_score in [0, 1]
 
-        # do update
-        loaded_default_tournament.update_current_round(new_res)
+    # def test_add_1st_results(self, four_players, last_tournament):
+    #     """add results for round 1"""
 
-        # next round
-        loaded_default_tournament._next_round()
+    #     # store current_round_number
+    #     current_round_number = last_tournament.current_round_number
 
-        # reload from the database the correct tournament
-        tournament_id = loaded_default_tournament.tournament_id
-        tournament = Tournament.search(tournament_id)
+    #     # use random results for the first round
+    #     new_res = [
+    #         # 1st match
+    #         (
+    #             # player 1 of 1st match
+    #             (four_players[0].player_id, 1),  # id, score
+    #             # player 2 of 1st match
+    #             (four_players[1].player_id, 0),  # id, score
+    #         ),
+    #         # 2nd match
+    #         (
+    #             # 1st player of 2nd match
+    #             (four_players[2].player_id, 1),  # id, score
+    #             # 2nd player of the 2nd match
+    #             (four_players[3].player_id, 0),  # id, score
+    #         ),
+    #     ]
 
-        # check current_round_number is OK
-        if tournament is not None:
-            assert (current_round_number + 1) == tournament.current_round_number
-        else:
-            logging.warning("Tournament is None")
+    #     # do update
+    #     last_tournament.update_current_round(new_res)
 
-    def test_add_3rd_results(self, four_players, loaded_default_tournament):
-        """add results for round 1"""
+    #     # next round
+    #     last_tournament._next_round()
 
-        abcd_list = [(0, 2, 1, 3), (0, 3, 1, 2)]
-        for a, b, c, d in abcd_list:
+    #     # reload from the database the correct tournament
+    #     tournament_id = last_tournament.tournament_id
+    #     tournament = Tournament.search(tournament_id)
 
-            # store current_round_number
-            current_round_number = loaded_default_tournament.current_round_number
+    #     # check current_round_number is OK
+    #     if tournament is not None:
+    #         assert (current_round_number + 1) == tournament.current_round_number
+    #     else:
+    #         logging.warning("Tournament is None")
 
-            # use random results for the first round
-            new_res = [
-                # 1st match
-                (
-                    # player 1 of 1st match
-                    (four_players[a].player_id, 1),  # id, score
-                    # player 2 of 1st match
-                    (four_players[b].player_id, 0),  # id, score
-                ),
-                # 2nd match
-                (
-                    # 1st player of 2nd match
-                    (four_players[c].player_id, 1),  # id, score
-                    # 2nd player of the 2nd match
-                    (four_players[d].player_id, 0),  # id, score
-                ),
-            ]
+    # def test_add_3rd_results(self, four_players, last_tournament):
+    #     """add results for round 1"""
 
-            # do update
-            loaded_default_tournament.update_current_round(new_res)
+    #     abcd_list = [(0, 2, 1, 3), (0, 3, 1, 2)]
+    #     for a, b, c, d in abcd_list:
 
-            # next round
-            loaded_default_tournament._next_round()
+    #         # store current_round_number
+    #         current_round_number = last_tournament.current_round_number
 
-            #
-            logging.critical(
-                f"loaded_default_tournament.tournament_id: {loaded_default_tournament.tournament_id}"
-            )
+    #         # use random results for the first round
+    #         new_res = [
+    #             # 1st match
+    #             (
+    #                 # player 1 of 1st match
+    #                 (four_players[a].player_id, 1),  # id, score
+    #                 # player 2 of 1st match
+    #                 (four_players[b].player_id, 0),  # id, score
+    #             ),
+    #             # 2nd match
+    #             (
+    #                 # 1st player of 2nd match
+    #                 (four_players[c].player_id, 1),  # id, score
+    #                 # 2nd player of the 2nd match
+    #                 (four_players[d].player_id, 0),  # id, score
+    #             ),
+    #         ]
 
-    def test_get_score(self, four_players, loaded_default_tournament):
-        """Test the get_score method"""
+    #         # do update
+    #         last_tournament.update_current_round(new_res)
 
-        # Iterate through each player
-        for player in four_players:
-            # Get the player's ID
-            player_id = player.player_id
+    #         # next round
+    #         last_tournament._next_round()
 
-            # Call the get_score method for the current player
-            player_score = loaded_default_tournament.get_score(player_id)
+    #         #
+    #         logging.critical(
+    #             f"last_tournament.tournament_id: {last_tournament.tournament_id}"
+    #         )
 
-            # Log the player's ID and score
-            logging.info(f"Player ID: {player_id}, Score: {player_score}")
+    # def test_get_score(self, four_players, last_tournament):
+    #     """Test the get_score method"""
 
-            # Check if the actual score is an integer
-            assert isinstance(player_score, int)
+    #     # Iterate through each player
+    #     for player in four_players:
+    #         # Get the player's ID
+    #         player_id = player.player_id
+
+    #         # Call the get_score method for the current player
+    #         player_score = last_tournament.get_score(player_id)
+
+    #         # Log the player's ID and score
+    #         logging.info(f"Player ID: {player_id}, Score: {player_score}")
+
+    #         # Check if the actual score is an integer
+    #         assert isinstance(player_score, int)