+++ CHANGELOG.md
@@ -1,4 +1,12 @@
 
+# 2024-03-22 - 10h00 - Code Review
+
+## Updated 
+- Add conftest
+- Add new 4 players 
+- rename fixture
+- change curent round number at 0 whe, status change to in progress
+
 # 2024-03-15 - 17h00 - Code Review